+++ UnitTests/Regressions/IssueReportsFixture.cs
@@ -1474,6 +1474,69 @@ namespace Moq.Tests.Regressions
 
 		#endregion
 
+		#region 438
+
+		public class Issue438
+		{
+			[Fact]
+			public void SetupAllPropertiesCanSetupSeveralSiblingPropertiesOfTheSameType()
+			{
+				var resultMock = new Mock<IResult> { DefaultValue = DefaultValue.Mock };
+				resultMock.SetupAllProperties();
+				var result = resultMock.Object;
+				result.Part1.Name = "Foo";
+				result.Part2.Name = "Bar";
+
+				Assert.Equal("Foo", result.Part1.Name);
+				Assert.Equal("Bar", result.Part2.Name);
+			}
+
+			public interface IResult
+			{
+				ISubResult Part1 { get; }
+				ISubResult Part2 { get; }
+			}
+
+			public interface ISubResult
+			{
+				string Name { get; set; }
+			}
+
+			[Fact]
+			public void SetupAllPropertiesCanDealWithSelfReferencingTypes()
+			{
+				var mock = new Mock<INode>() { DefaultValue = DefaultValue.Mock };
+				mock.SetupAllProperties();
+				Assert.Null(mock.Object.Parent);
+			}
+
+			public interface INode
+			{
+				INode Parent { get; }
+			}
+
+			[Fact]
+			public void SetupAllPropertiesCanDealWithMutuallyReferencingTypes()
+			{
+				var mock = new Mock<IPing>() { DefaultValue = DefaultValue.Mock };
+				mock.SetupAllProperties();
+				Assert.NotNull(mock.Object.Pong);
+				Assert.Null(mock.Object.Pong.Ping);
+			}
+
+			public interface IPing
+			{
+				IPong Pong { get; }
+			}
+
+			public interface IPong
+			{
+				IPing Ping { get; }
+			}
+		}
+
+		#endregion
+
 		// Old @ Google Code
 
 		#region #47